fix: handle empty fields as empty string
By default empty fields were being converted to 0, and trailing empty
fields were not added to the row. This was in part due to the default
reviver number conversion. The reviver parameter makes it easy to do
this, so let the user handle instead of doing it by default.

BREAKING CHANGE: Numeric text is no longer converted to numbers.

If you need this, you will need to provide a reviver function to do
the parsing.